--- ChristerData.csv
@@ -1,37 +1,0 @@
-,Christer
-0,360.0
-1,650.0
-2,450.0
-3,725.0
-4,540.0
-5,450.0
-6,780.0
-7,0.0
-8,450.0
-9,0.0
-10,0.0
-11,0.0
-12,270.0
-13,400.0
-14,600.0
-15,550.0
-16,700.0
-17,440.0
-18,500.0
-19,700.0
-20,600.0
-21,700.0
-22,0.0
-23,360.0
-24,480.0
-25,650.0
-26,600.0
-27,600.0
-28,0.0
-29,270.0
-30,270.0
-31,0.0
-32,0.0
-33,480.0
-34,480.0
-35,360.0
