--- Data.csv
@@ -1,37 +1,0 @@
-,Training load
-0,
-1,180.0
-2,510.0
-3,780.0
-4,330.0
-5,720.0
-6,600.0
-7,660.0
-8,0.0
-9,720.0
-10,216.0
-11,0.0
-12,515.0
-13,0.0
-14,0.0
-15,0.0
-16,1120.0
-17,480.0
-18,0.0
-19,1080.0
-20,1500.0
-21,0.0
-22,0.0
-23,480.0
-24,0.0
-25,0.0
-26,721.0
-27,1260.0
-28,532.0
-29,240.0
-30,
-31,680.0
-32,0.0
-33,720.0
-34,
-35,0.0
