--- EinarTL.csv
@@ -1,36 +1,0 @@
-,0
-0,180.0
-1,510.0
-2,780.0
-3,330.0
-4,720.0
-5,600.0
-6,660.0
-7,0.0
-8,720.0
-9,216.0
-10,0.0
-11,515.0
-12,0.0
-13,0.0
-14,0.0
-15,1120.0
-16,480.0
-17,0.0
-18,1080.0
-19,1500.0
-20,0.0
-21,0.0
-22,480.0
-23,0.0
-24,0.0
-25,721.0
-26,1260.0
-27,532.0
-28,240.0
-29,
-30,680.0
-31,0.0
-32,720.0
-33,
-34,0.0
