--- FilipData.csv
@@ -1,42 +1,0 @@
-,Filip
-0,609.0
-1,540.0
-2,330.0
-3,481.0
-4,212.0
-5,402.0
-6,520.0
-7,0.0
-8,616.0
-9,48.0
-10,464.0
-11,120.0
-12,585.0
-13,695.0
-14,60.0
-15,304.0
-16,300.0
-17,352.0
-18,370.0
-19,576.0
-20,120.0
-21,544.0
-22,0.0
-23,210.0
-24,272.0
-25,0.0
-26,150.0
-27,69.0
-28,310.0
-29,417.0
-30,375.0
-31,816.0
-32,170.0
-33,336.0
-34,268.0
-35,330.0
-36,1065.0
-37,0.0
-38,468.0
-39,152.0
-40,0.0
