--- JessicaData.csv
@@ -1,35 +1,0 @@
-,Jessica
-0,1008.0
-1,0.0
-2,0.0
-3,535.0
-4,154.0
-5,0.0
-6,455.0
-7,0.0
-8,0.0
-9,0.0
-10,725.0
-11,360.0
-12,762.0
-13,0.0
-14,1008.0
-15,0.0
-16,0.0
-17,705.0
-18,0.0
-19,0.0
-20,240.0
-21,0.0
-22,1305.0
-23,0.0
-24,544.0
-25,475.0
-26,680.0
-27,0.0
-28,0.0
-29,360.0
-30,0.0
-31,640.0
-32,0.0
-33,460.0
