--- JonnyData.csv
@@ -1,38 +1,0 @@
-,Jonny
-0,0.0
-1,464.0
-2,0.0
-3,351.0
-4,315.0
-5,448.0
-6,536.0
-7,536.0
-8,524.0
-9,1176.0
-10,436.0
-11,770.0
-12,440.0
-13,456.0
-14,0.0
-15,714.0
-16,480.0
-17,385.0
-18,480.0
-19,432.0
-20,575.0
-21,0.0
-22,376.0
-23,824.0
-24,428.0
-25,0.0
-26,512.0
-27,432.0
-28,412.0
-29,360.0
-30,408.0
-31,816.0
-32,600.0
-33,360.0
-34,340.0
-35,600.0
-36,472.0
