--- ThomasData.csv
@@ -1,38 +1,0 @@
-,Thomas
-0,900.0
-1,270.0
-2,360.0
-3,0.0
-4,296.0
-5,399.0
-6,333.0
-7,0.0
-8,525.0
-9,0.0
-10,315.0
-11,324.0
-12,0.0
-13,240.0
-14,945.0
-15,440.0
-16,0.0
-17,0.0
-18,315.0
-19,0.0
-20,240.0
-21,0.0
-22,0.0
-23,0.0
-24,370.0
-25,558.0
-26,490.0
-27,0.0
-28,0.0
-29,360.0
-30,300.0
-31,385.0
-32,490.0
-33,536.0
-34,455.0
-35,0.0
-36,600.0
