--- extra.py
@@ -1,50 +1,0 @@
-# for x in range(len(nameWorkout)):
-#     if nameWorkout[x] == 'C':
-#         ChristerTL.append(duration[x] * intensity[x])
-#         hrv_value = check_hrv(dateWorkout[x], dateHRV, nameWorkout[x], nameHRV, HRV)
-#         ChristerHRV.append(hrv_value)
-#     elif nameWorkout[x] == 'Einar Ingemarsson ':
-#         EinarTL.append(duration[x] * intensity[x])
-#         hrv_value = check_hrv(dateWorkout[x], dateHRV, nameWorkout[x], nameHRV, HRV)
-#         EinarHRV.append(hrv_value)
-#     elif nameWorkout[x] == 'Thomas':
-#         ThomasTL.append(duration[x] * intensity[x])
-#         hrv_value = check_hrv(dateWorkout[x], dateHRV, nameWorkout[x], nameHRV, HRV)
-#         ThomasHRV.append(hrv_value)
-#     elif nameWorkout[x] == 'Jessica':
-#         JessicaTL.append(duration[x] * intensity[x])
-#         hrv_value = check_hrv(dateWorkout[x], dateHRV, nameWorkout[x], nameHRV, HRV)
-#         JessicaHRV.append(hrv_value)
-#     elif nameWorkout[x] == 'Filip Helmroth':
-#         FilipTL.append(duration[x] * intensity[x])
-#         hrv_value = check_hrv(dateWorkout[x], dateHRV, nameWorkout[x], nameHRV, HRV)
-#         FilipHRV.append(hrv_value)
-#     elif nameWorkout[x] == 'Jonny':
-#         JonnyTL.append(duration[x] * intensity[x])
-#         hrv_value = check_hrv(dateWorkout[x], dateHRV, nameWorkout[x], nameHRV, HRV)
-#         JonnyHRV.append(hrv_value)
-
-
-# print(dateWorkout[50:54])
-# print(nameWorkout[50:54])
-#
-#
-# print("Date: ", dateHRV[49])
-# print("Name: ", nameHRV[49])
-# print("HRV: ", HRV[49])
-#
-#
-# print("nameWorkout: ", nameWorkout[50])
-# print("dateWorkout: ", dateWorkout[50])
-#
-# print(intensity[50:54])
-# print(duration[50:54])
-# print(check_workout(dateWorkout, nameWorkout, dateHRV[49], nameHRV[49], intensity, duration))
-
-# print("intensity: ", intensity[tk])
-                # print("duration: ", duration[tk])
-
-
-# for list, list_ in zip(newHRV, newTL):
-#     list_.insert(0, 0)
-#     list.append(0)
